refactor(core): rename ticket model filename in preparation for base ticket model
ref: #724 #723
