+++ app/core/tests/unit/ticket_depreciated/test_ticket_depreciated_common.py
