+++ app/core/tests/unit/ticket_depreciated/test_ticket_depreciated_views.py
