+++ app/itim/serializers/change.py
@@ -4,7 +4,7 @@ from rest_framework.fields import empty
 from app.serializers.user import UserBaseSerializer
 
 from core.models.ticket.ticket_category import TicketCategory
-from core.serializers.ticket import (
+from core.serializers.ticket_depreciated import (
     Ticket,
     TicketBaseSerializer,
     TicketModelSerializer,
