+++ app/itim/serializers/problem.py
@@ -3,7 +3,7 @@ from rest_framework import serializers
 from app.serializers.user import UserBaseSerializer
 
 from core.models.ticket.ticket_category import TicketCategory
-from core.serializers.ticket import (
+from core.serializers.ticket_depreciated import (
     Ticket,
     TicketBaseSerializer,
     TicketModelSerializer,
