+++ app/itim/viewsets/incident.py
@@ -13,7 +13,7 @@ from itim.serializers.incident import (
     IncidentTicketViewSerializer,
 )
 
-from core.viewsets.ticket import TicketViewSet
+from core.viewsets.ticket_depreciated import TicketViewSet
