+++ app/project_management/viewsets/project_task.py
@@ -14,7 +14,7 @@ from project_management.serializers.project_task import (
     ProjectTaskTicketViewSerializer,
 )
 
-from core.viewsets.ticket import TicketViewSet
+from core.viewsets.ticket_depreciated import TicketViewSet
